frontend/pyproject.toml
[tool.poetry]
name = "rag-front-streamlit"
version = "0.1.0"
description = ""
authors = ["Federico Castañeda <fedecasta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
streamlit = "^1.32.2"
watchdog = "^4.0.2"
langchain-community = "^0.2.12"
langchain-openai = "^0.1.22"
faiss-cpu = "^1.8.0.post1"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
langchain-pinecone = "^0.1.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
